Moving default story icon handling to experiments directory
